public: Added QueryProvider
This was necessary because each new instance of SearchBar would
re-create the query state, which would cause it to be different
between new instances.

This also helps us to manage the query in a global area. This means
that we can type a query in the search bar on the **Home** page and
have the same query populate in the search bar on the **results** page.

Yes, we _could_ just get the query from the URL, but having a
central management component is very useful.