public fix: Results are now reset on new queries
The skeleton component never loaded twice because it expects the
results to be undefined. The solution, of course, is to reset the
Results array whenever the query changes.

To prevent duplicate queries, I added a new usePrevious hook which
will compare the current query to the previous one, and only allow
requests if the query is different.