public: Better results!
I was wondering why the results Gofë was returning were crappy. Then it
occurred to me that we may not be seeing the first page of the results.
Indeed, it turns out that we have been viewing page 2 for a little
while.

This was due to the fact that I thought pages would be indexed at 1. I
asked to confirm this, but you may not have seen it.

This commit fixes this issue by subtracting 1 from the p value on every query to the Go
API. I did this because we want people to view "friendly" page numbers,
with the first one being 1. Therefore, the NJS data and URL all work
with +1 the actual index value.

When you visit ?p=4, it is requesting the page from the Go API with
?p=3.