public implement #18: Related searches
This just adds the suggestions to the bottom of the page (Google does
them above pagination, so I did the same).

When you click a suggestion, the SearchBar suggestions are shown, which
highlights, again, issue #15. Since we've decided to put that on hold
for the time being, I figured you'd be okay to tolerate this minor
    inconsistency.