public: Improve error handling
This introduces a ton of new error handling improvements—both in terms of the requests and the components.

One should note that the `errorType` you see throughout is only related to _request_ errors, not undefined or empty values.

While the suggestions API does indeed have a way to check for errors (given by the `status` export), I decided not to implement anything related to it this time because the SearchBar is not yet done being refactored. Right now, it's fine. Later, we'll just add some error handling support to the refactor. Easy.

My other changes are summarized below:

- I finally removed the carets from the package.json versions and recompiled the yarn.lock file. All is well.
- I added a `repository` attribute to `config.json`. Should we add this to .gitignore and just update a `config.json.local`? After all, the VPS will need different stuff, right?
- I added a dev package: `@types/react-keyboard-event-handler`. This just fixes a type error in the Keyboard component.
- I also added my own library, `@infinium/hydro` to add some UI elements. This will be much more useful once we add more features (like settings, etc.).
- New `SearchError.tsx` component which handles the primary error view for Search results. It changes the content depending on the type of request error.
- I added a 'Report a bug' link in the footer, since it is related to errors.
- New 'x' icon
- New `err.ts` file which contains modular error-handling functions related specifically to requests. All errors relating to empty or undefined data is within components.
- Because of the new request error-handling functions, I was able to greatly reduce the amount of boilerplate code in the `services/*` files. It is much simpler now.
- Various error-related changes. See commit.