public: Add support for different themes!
This is a rather large update by the quantity of changes. This is mostly due to how I was previously handling the color system. It was rather easy, you see: Jupiter has CSS theme values that toggle between dark and light. All I had to do was install `next-themes` and use the CSS vars like `ui-2`.

This worked for the time when we didn't want extra customization. Now, everything has changed.

Instead of relying on the Jupiter-defined values, we create new ones and attach them to all areas in the site. Then, we simply change the class on the body to toggle these colors/customizations.

This means that we no longer have a default "dark/light mode." `next-themes` was removed. Instead, we set the theme to a default of `auto` which toggles between `modern-light` and `modern-dark` depending on the `prefer-color-scheme` CSS value.

In other words, by default, it will set the theme to the system.

You can toggle all of this by using the settings button in the top-right; this opens a new settings modal that will allow us to add more options in the future.

The actual changes to the code are largely irrelevant. It was just a good amount of changes; that's CSS after all.