public: Fix suggestions direct mutation issue
We can't ever reassign state objects manually. That may appear to work, but we'll get some nasty issues later on. We must always work on a new reference of the data or use the React-given functional updaters.

In this case, filter() and map() return a _new_ array, so it's fine.