public: Improved the reliability of useSettings
This patch mostly improves the error-handling of the settings on the initial read. It now checks the default settings and ensures that all stored settings have valid (and approved) values. If they don't have an approved value or if the setting itself doesn't exist, it is reset to the default or deleted, respectively.