public: Fix suggestions on homepage load
On the homepage, we autoFocus the SearchBar. Since the suggestions being
shown isn't affected by the query itself, we have to show the
suggestions automatically.

Since I didn't want it to always show "No suggestions" by default, I
also added some logic to show "Type to get suggestions" on the first
load, just to be more friendly about it. Of course, if you type a query
that is too specific, it'll go back to the default "No suggestions."