api: 'unescape' Google's URLs
Google does not link directly to the result URLs. Instead, it uses query
parameters to let itself know where the user is going, and then
redirects the user to that page:

`<a href='/url?q=https://idioms.thefreedictionary.com/pass%2Bmuster'>`

Here, %2B is in reality '+'. This was causing a host of issues on our
end. This commit therefore fixes this: we now decode the encoded URLs.