public fix: Manually set .s-context BG color
I left the skeleton BG for the context text as the default since it
appeared to work fine. Indeed, it did, for all themes but one: Dracula.

I suspect it is just due to the way Dracula determines its colors (I
copied the BG colors directly from the theme devs themselves instead of
determining the shading of the BG manually like I had done with every
other theme.)

Anyhow, this makes little visible difference (it defaults, I think, to a
gradient background of ui-1 and ui-2).