public: Beautify URLs
At a high level, this commit takes the URL of every result, splits it
into a protocol, host, and pathname (pathname is split, too) and places
rightChevron icons beside them. This makes the URLs much easier to
digest and, in my opinion, makes everything so much better.

Throughout the code, you'll notice a few comments related to the
functionality here. This was an... open-ended feature and the way I took
it may not be ideal; clearly, however, I realized this and have
identified those areas where functionality can be altered to tweak the
URL output.

---

Right now, it defaults to a 25 character length on the host and paths
before truncating them. It also slices the paths to a max of 3, as that
is about how much space we can handle.

The `determinePathTrimLength` function is important here, as it dictates
how long each path should be. Essentially, if the previous item (`host`
as `0`) is longer than 25, it will trim the current path to 8.

In practice, this means that if we have a really long URL like
https://999999999999999999999999999999.com/999999999999999999999999999999/999999999999999999999999999999/999999999999999999999999999999,
it will only show the very second path as a 25 character trim. This
means no URL is too long to handle, so we won't get any weird layout
errors.

My focus here was precisely that: to handle any length of URL while
still displaying it "nicely." As such, the functionality of certain
functions may not be focused towards informational output, but visual
output.

As a fallback to all of this, I added the title="" attribute to the link
so that one can always see the full URL.

To do this, I had to modify ExternalLink to accept the standard
React.AnchorHTMLAttributes<any> props to allow the title attribute
implicitly.

In addition to these changes, I added `word-break: break-all;` to the
Description and Context, so that we won't get weird layout issues from
those, either.