public: Make RelatedSearches use staticSuggestions
It was _really_ bothering me that the RelatedSearches in the Results
area would be the same as the suggestions in the SearchBar. I understand
how it can be nice and maybe even convenient, but it's not normal
behavior.

I find more utility in retaining the old suggestions until a new search
is executed.