public: Fix RelatedSearches error; etc.
I also renamed the QueryProvider to SearchProvider as that more closely
defines its purpose (as of now, and as of later).

This required various import changes, etc.

The issue with the RelatedSearches was that it wasn't showing on a fresh
load because the suggestions promise wasn't yet fulfilled.

To get around this, I had to extract the suggestions logic to a separate
function which is exported. The function depends on the query parameter
which allows the RelatedSearches area to make its own request based on
the staticQuery.

I figured, "what's one extra suggestions request on a query that rarely
changes?"

If you disagree, let me know.