public: Fix results overflow on mobile
This happened because the URL was allowed to expand beyond what a mobile
device is capable of showing.

Due to your aversion to dependencies, I've done a really shitty
monkey-patch to hide the last two items (icon, span) on mobile. We're
going to hope that this will suffice for the time being.

The alternative is to reduce the font-size of the URL on mobile, which
is already small enough.

The ideal solution would be to use a dependency (preferably
react-device-detect) to allow us to programmatically modify the length
of the path trim depending on the size of the device.

Consider it.