public: <View> now wraps <RootLayout>
I just made a minor mistake in the organization of the _app components.

<RootLayout> can (does) change for each page, which _was_ problematic
when it wrapped the <View> as well. This is because <View> contains the
DOM-related settings logic.

In effect, this meant that, on occasion and without any discernable
pattern, the home page and the search results page would be
preferentially out of sync (so to speak).