api: improve sanitize()
I was under the impression that query parameters such as 'sa' and 'ved'
were for the destination websites. However, these parameters are for
Google. As such, the 'pure' website link is stored in the query
parameter 'q': we just need to extract this.

Also, if there's any trouble parsing the URL, it is now ignored.