public: Unrestrict loading of Search results
Previously, the conditions in the Search page would basically show a
completely different component depending on if the results were loading
or if no results were found.

As a consequence, this prevented other things on the page (mostly
RelatedSearches) to be delayed. They would only be rendered after the
search results came in.

This commit "fixes" this by moving the conditions to within the primary
return value—allowing the Search results and suggestions to be fetched
simultaneously.

Various minimal changes were necessary, but this was rather
straightforward.