public: Improve SettingsModal
I added a new CheckboxSetting component which is clearly more applicable
for certain settings. It also adds a much cleaner contrast and brings
    the whole thing together.

I had considered placing the checkboxes at the top of the modal, but I
figured that the theme setting is probably the most important. If you
wanna switch it, lmk.