public: Redesign Settings component
This commit moves away from the Modal component exported by Hydro.
Instead, I've decided to recreate it from scratch.

It's no longer a modal, but a little panel that's positioned relative to
the settings button. As such, it should be much less intrusive.

In addition to this, this also allows us to make more fine-tuned
customizations to the settings as we start adding more of them.