Add Notice feature (#62)
This allows us to create arbitrary notices to our end users. I began with a simple one demonstrating our theme feature.

---

The code is "smart" such that whenever a Notice is rendered, it automatically cleans up the users `localStorage` by removing any items relating to previous notices when there is another active notice.

It also doesn't have any perceptible (in my tests) rendering issue. By default, no notice is rendered; this is changed only if the `localStorage` item is missing or if it's set to `true`.

---

Co-authored-by: Tristan B <tristan@infinium.earth>
Reviewed-on: https://codeberg.org/ar324/gofe/pulls/62